break out and improve instantiate & pool tests
