+++ artifacts/checksums.txt
@@ -1,3 +1,3 @@
-7f0d72bacb0ab8b05ae84f08ed0a3aa0c72807a7d3beccaad4a891ab9e5fabe7  dca.wasm
+592af99c568eb022bc7982f63278a78ec3e33c88f2b6bfee5124ce7b176849fd  dca.wasm
 aaf49661cb706442ea8bc66e61aab3beaf3c9aa1552465dbad7617642736b44d  fin.wasm
 c37975e297930feed7dbe11d02ac589c3b965db4ee897981b73d16441273d755  staking_router.wasm
